Refactor product pricing components to use VariantPrices for consistent price display
