--- app/components/PaginatedResourceSection.tsx
@@ -1,41 +1,0 @@
-import { Pagination } from "@shopify/hydrogen";
-import type * as React from "react";
-
-/**
- * <PaginatedResourceSection > is a component that encapsulate how the previous and next behaviors throughout your application.
- */
-export function PaginatedResourceSection<NodesType>({
-  connection,
-  children,
-  resourcesClassName,
-}: {
-  connection: React.ComponentProps<typeof Pagination<NodesType>>["connection"];
-  children: (props: { node: NodesType; index: number }) => React.ReactNode;
-  resourcesClassName?: string;
-}) {
-  return (
-    <Pagination connection={connection}>
-      {({ nodes, isLoading, PreviousLink, NextLink }) => {
-        const resourcesMarkup = nodes.map((node, index) =>
-          children({ node, index }),
-        );
-
-        return (
-          <div>
-            <PreviousLink>
-              {isLoading ? "Loading..." : <span>↑ Load previous</span>}
-            </PreviousLink>
-            {resourcesClassName ? (
-              <div className={resourcesClassName}>{resourcesMarkup}</div>
-            ) : (
-              resourcesMarkup
-            )}
-            <NextLink>
-              {isLoading ? "Loading..." : <span>Load more ↓</span>}
-            </NextLink>
-          </div>
-        );
-      }}
-    </Pagination>
-  );
-}
