--- app/components/compare-at-price.tsx
@@ -1,20 +1,0 @@
-import { useMoney } from "@shopify/hydrogen";
-import type { MoneyV2 } from "@shopify/hydrogen/storefront-api-types";
-import { cn } from "~/utils/cn";
-
-export function CompareAtPrice({
-  data,
-  className,
-}: {
-  data: MoneyV2;
-  className?: string;
-}) {
-  const { currencyNarrowSymbol, withoutTrailingZerosAndCurrency } =
-    useMoney(data);
-  return (
-    <span className={cn("strike text-(--color-compare-price-text)", className)}>
-      {currencyNarrowSymbol}
-      {withoutTrailingZerosAndCurrency}
-    </span>
-  );
-}
