+++ app/components/product/quick-shop.tsx
@@ -20,9 +20,9 @@ import { AddToCartButton } from "~/components/product/add-to-cart-button";
 import { ProductMedia } from "~/components/product/product-media";
 import { Quantity } from "~/components/product/quantity";
 import { Skeleton } from "~/components/skeleton";
-import { VariantPrices } from "~/components/variant-prices";
 import type { ProductData } from "~/routes/($locale).api.product";
 import { ProductBadges } from "./badges";
+import { VariantPrices } from "./variant-prices";
 import { VariantSelector } from "./variant-selector";
 
 interface QuickViewData {
